requirements.txt
numpy
scipy
scikit-learn==1.2.2
matplotlib
pandas
fastapi
uvicorn
gunicorn==19.9.0
uvloop
httptools
requests
jinja2
paho-mqtt